Issue #2807: Reorganize subcomponents & tests used only by ExperimentPage
