+++ frontend/src/app/containers/ExperimentPage/index.js
@@ -10,13 +10,9 @@ import { buildSurveyURL, experimentL10nId, formatDate } from '../../lib/utils';
 import NotFoundPage from '../NotFoundPage';
 
 import EmailDialog from '../../components/EmailDialog';
-import ExperimentDisableDialog from '../../components/ExperimentDisableDialog';
-import ExperimentEolDialog from '../../components/ExperimentEolDialog';
-import ExperimentTourDialog from '../../components/ExperimentTourDialog';
 import GraduatedNotice from '../../components/GraduatedNotice';
 import LocalizedHtml from '../../components/LocalizedHtml';
 import ExperimentCardList from '../../components/ExperimentCardList';
-import ExperimentPreFeedbackDialog from '../../components/ExperimentPreFeedbackDialog';
 import View from '../../components/View';
 import Warning from '../../components/Warning';
 
@@ -24,6 +20,10 @@ import ExperimentPlatforms from '../../components/ExperimentPlatforms';
 import Banner from '../../components/Banner';
 import LayoutWrapper from '../../components/LayoutWrapper';
 
+import ExperimentPreFeedbackDialog from './ExperimentPreFeedbackDialog';
+import ExperimentDisableDialog from './ExperimentDisableDialog';
+import ExperimentEolDialog from './ExperimentEolDialog';
+import ExperimentTourDialog from './ExperimentTourDialog';
 import IncompatibleAddons from './IncompatibleAddons';
 import TestpilotPromo from './TestpilotPromo';
